Converted full screenshot to used RAM based storage
Initially, the NZA program stored a screenshot of the users screen on 
the disk, however, in some circumstances storing this screenshot on the 
disk was impossible due to insufficent permissions to write to the disk. 
This commit remedies this issue by storing this screenshot entirely in 
RAM thus absolving the need to store the screenshot on the disk.